backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic[email]==2.10.3
pydantic-settings==2.6.1
email-validator>=2.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.17
httpx==0.28.1
anthropic>=0.45.2
apscheduler==3.10.4
langgraph
langchain-core
langchain-anthropic
langchain-ollama>=0.2.0
openai>=1.0.0
openpyxl>=3.1.0
